Save Study
Save Study feature implementation need to review

+++ GoalManager/app/src/main/java/com/androidvoyage/goalmanager/datamodels/Post.kt
@@ -16,7 +16,7 @@ class Post : Serializable{
     }
 
     @PrimaryKey(autoGenerate = true)
-    var postId : Int = 0
+    var postId : Long = 0
 
     @ColumnInfo(name ="Header")
     var header : String = ""
